% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_embedding.R
\name{bcp_coal}
\alias{bcp_coal}
\title{bcp_coal}
\usage{
bcp_coal(n)
}
\arguments{
\item{n}{the number of lineages at the start of the process}
}
\value{
a list, the first element is the state space matrix,
the second the transition probability matrix
}
\description{
Computes state space and transition probability matrix for the
Block Counting process
}
