% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{BCPssp}
\alias{BCPssp}
\title{BCP.ssp}
\format{
a vector of length 25.
}
\source{
Hard coded
}
\usage{
BCPssp
}
\description{
The size of state spaces for the block counting process for n ranging
between 1 and 25.
}
\details{
the n'th entry is the number of states in BCP with n lineages.
}
\keyword{datasets}
