% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetupPreloads.R
\name{binary_matrix}
\alias{binary_matrix}
\title{binary_matrix}
\usage{
binary_matrix(d)
}
\arguments{
\item{d}{Maximum number of digits of binary number}
}
\value{
Matrix of binary numbers
}
\description{
Returns a matrix of binary numbers up to a specified number of digits
}
