% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPHFunctions.R
\name{binned_probs_E}
\alias{binned_probs_E}
\title{binned_probs_E}
\usage{
binned_probs_E(m, Probs)
}
\arguments{
\item{m}{number of observations in sample}

\item{Probs}{distribution of E}
}
\value{
A list of the bins as well as the expected number in each bin
}
\description{
Bins the probabilities computed above. Used in the GoF test
}
