% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityAndPathFunctions.R
\name{check_path}
\alias{check_path}
\title{check_path}
\usage{
check_path(path, TPM)
}
\arguments{
\item{path}{a vector of length n-1}

\item{TPM}{a transition probability matrix}
}
\value{
TRUE/FALSE depending on whether the path is possible
}
\description{
Checks if a path has probability > 0
}
