% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityAndPathFunctions.R
\name{checkN}
\alias{checkN}
\title{checkN}
\usage{
checkN(n, lower = 3, upper = 27)
}
\arguments{
\item{n}{the input}

\item{lower}{the lower bound of the input}

\item{upper}{the upper bound of the input}
}
\value{
integer n or raises Error
}
\description{
Checks input for being an integer between upper and lower bounds
}
