% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPHFunctions.R
\name{covariance_matrix_F}
\alias{covariance_matrix_F}
\title{covariance_matrix_F}
\usage{
covariance_matrix_F(LeftProducts, RightProducts, IndexList, RewardMatrixF)
}
\arguments{
\item{LeftProducts}{A list of the products init_dist * U * diag(r_{i,j})
where (i,j) is a non-fixed index. Computed via left_products_seq}

\item{RightProducts}{A list of the products U * diag(r_{i,j}) * e
where (i,j) is a non-fixed index. Computed via right_products_seq}

\item{IndexList}{A list of all non fixed indices. Made using non_fixed_indices}

\item{RewardMatrixF}{reward matrix of the non-fixed entries. Can be computed
using reward_matrix_F.}
}
\value{
the covariance matrix of all non fixed entries
}
\description{
Computes the covariance matrix of the non fixed F-matrix indices.
}
