% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPHFunctions.R
\name{covariance_matrix_SE}
\alias{covariance_matrix_SE}
\title{covariance_matrix_SE}
\usage{
covariance_matrix_SE(CovMatNoFixed, IndexList)
}
\arguments{
\item{CovMatNoFixed}{the covariance matrix of all non fixed entries}

\item{IndexList}{A list of all non fixed indices. Made using non_fixed_indices}
}
\value{
the covaraince matrix of the external branch length and sum of
non-fixed entries
}
\description{
Computes the 2 x 2 covariance matrix of the external branch length and sum
of non-fixed entries.
}
