% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityAndPathFunctions.R
\name{DfromF}
\alias{DfromF}
\title{DfromF}
\usage{
DfromF(Fmat)
}
\arguments{
\item{Fmat}{Valid F-Matrix}
}
\value{
D-Matrix corresponding to the same ranked unlabelled tree
}
\description{
Transforms an F-Matrix into a D-matrix
}
