% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetupPreloads.R
\name{diffM25_from_binM25}
\alias{diffM25_from_binM25}
\title{diffM25_from_binM25}
\usage{
diffM25_from_binM25(binM25)
}
\arguments{
\item{binM25}{a matrix of all binary numbers that can be represented
with 25 digits. Computed using binary_matrix(25). Alternatively can be loaded.}
}
\value{
the diffM25 matrix.
}
\description{
function computes the diffM25 matrix from the binM25 matrix.
}
