% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPHFunctions.R
\name{dist_E}
\alias{dist_E}
\title{dist_E}
\usage{
dist_E(TPME, InitDist, MinE, MaxE)
}
\arguments{
\item{TPME}{Sub-TPM of E. Can be computed using PhaseTypeR}

\item{InitDist}{initial distribution of E. Can be computed using PhaseTypeR}

\item{MinE}{the minimum external branch length}

\item{MaxE}{the maximum external branch length}
}
\value{
A vector of length max_E - min_E, with entry i = P(E = min_E - 1 + i)
}
\description{
Computes the probabilities P(E = k) for all k between min_E and max_E
}
