% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPHFunctions.R
\name{DPH_E}
\alias{DPH_E}
\title{DPH_E}
\usage{
DPH_E(BCP, RewardE)
}
\arguments{
\item{BCP}{An object of the discrete_phase_type class. Computed using
the DPH function from PhaseTypeR}

\item{RewardE}{The reward transform of the external branch length. Computed
using reward_E_BCP}
}
\value{
The sub-TPM and initial distribution of E
}
\description{
Computes the sub-TPM and initial distribution of E
}
