% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityAndPathFunctions.R
\name{external_branch_length}
\alias{external_branch_length}
\title{external_branch_length}
\usage{
external_branch_length(Fmat)
}
\arguments{
\item{Fmat}{An (n-1) x (n-1) F-matrix}
}
\value{
the external branch length of the ranked and unlabeled tree.
}
\description{
computes the external branch length of tree based on its F-matrix
}
