% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{ExternalBranchLengthDist25}
\alias{ExternalBranchLengthDist25}
\title{PMF of external branch length under the Kingman for n = 25
A precomputed vector of probabilities under the Kingman coalescent.}
\format{
A vector of length 133
}
\source{
Generated using dist_E
}
\usage{
ExternalBranchLengthDist25
}
\description{
This vector is used for statstical inference and hypothesis testing
in the RankedCoalescent package.
}
\details{
The vector represents the PMF with entry i = P(E = minE - 1 + i)
}
\keyword{datasets}
