% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityAndPathFunctions.R
\name{F_no_fixed}
\alias{F_no_fixed}
\title{F_no_fixed}
\usage{
F_no_fixed(Fmat, IndexList)
}
\arguments{
\item{Fmat}{An (n-1) x (n-1) F-matrix}

\item{IndexList}{A list of all non fixed indices.
Made using \code{non_fixed_indices}}
}
\value{
A vector of the non fixed entries
}
\description{
computes a vector of all non fixed entries in a given F-matrix
}
