% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityAndPathFunctions.R
\name{FfromD}
\alias{FfromD}
\title{FfromD}
\usage{
FfromD(Dmat)
}
\arguments{
\item{Dmat}{Valid D-Matrix}
}
\value{
F-Matrix corresponding to the same ranked unlabelled tree
}
\description{
Transforms a D-Matrix into an F-matrix
}
