% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityAndPathFunctions.R
\name{fibP1}
\alias{fibP1}
\title{fibP1}
\usage{
fibP1(n)
}
\arguments{
\item{n}{The index of the fibonacci number -1}
}
\value{
Fib(n+1)-1
}
\description{
Gives the (n+1)'th Fibonacci number minus 1
}
