% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityAndPathFunctions.R
\name{Fmat_to_path}
\alias{Fmat_to_path}
\title{Fmat_to_path}
\usage{
Fmat_to_path(Fmat, StSpM, Tiers)
}
\arguments{
\item{Fmat}{an (n-1) x (n-1) F-matrix}

\item{StSpM}{matrix of all states}

\item{Tiers}{A list of tiers made using make_tiers. Added for efficiency}
}
\value{
ordered path
}
\description{
Translates an F-matrix to its corresponding ordered path
}
