% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityAndPathFunctions.R
\name{FtoBlock}
\alias{FtoBlock}
\title{FtoBlock}
\usage{
FtoBlock(Fmat)
}
\arguments{
\item{Fmat}{Valid F-matrix}
}
\value{
Block Counting Matrix
}
\description{
Transforms an F-matrix into a representation of the block-counting process (giving SFS of each coalescence event)
}
