% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleFromBF.R
\name{gen_Fmats_from_BF}
\alias{gen_Fmats_from_BF}
\title{gen_Fmats_from_BF}
\usage{
gen_Fmats_from_BF(b, m = 1000, n = 25)
}
\arguments{
\item{b}{the parameter beta in the Blum-Francois model}

\item{m}{the size of sample. 1000 by default}

\item{n}{the size of the trees (number of leaves). 25 by default}
}
\value{
a sample of m F-matrices sampled from the Blum-Francois model
}
\description{
Samples F-matrices from the Blum-Francois model. The code is taken from
the f-matrix package, but changed so that it outputs in the desired format.
}
