% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetupPreloads.R
\name{generate_binary_files}
\alias{generate_binary_files}
\title{generate_binary_files}
\usage{
generate_binary_files()
}
\description{
Function creates three data files "binM25, diffM25, ndiff25". Needed to gene-
rate the transition probability matrix and state space using ranked_coal.
Preferably, read the binM25.rds file and apply the diffM25_from_binM25 and
ndiff25_from_binM25 functions to it.
}
