% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityAndPathFunctions.R
\name{getAdditiveDiff}
\alias{getAdditiveDiff}
\title{getAdditiveDiff}
\usage{
getAdditiveDiff(sdiag, sdVal)
}
\arguments{
\item{sdiag}{the index of the subdiagonal}

\item{sdVal}{the value of the subdiagonal}
}
\value{
Additive vectors that can be used to create new states
}
\description{
For a certain tier, gives the required differenceVectors
}
