% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticalTests.R
\name{GoF_E}
\alias{GoF_E}
\title{GoF_E}
\usage{
GoF_E(Fmats, Bins, ExpectedBinned, MinE, MaxE)
}
\arguments{
\item{Fmats}{a list of (n-1) x (n-1) F-matrices}

\item{Bins}{The bins needed to compute the GoF test statistics. Made using
make_bins function.}

\item{ExpectedBinned}{the expected number of observation within each bin
under the null hypothesis. Can be computed using the make_bins function}

\item{MinE}{the minimum external branch length}

\item{MaxE}{the maximum external branch length}
}
\value{
the p-value of the GoF test
}
\description{
Performs the goodnees of fit test based on the balance index E as described
in Theorem 6.1 in NAME OF PAPER.
}
