% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPHFunctions.R
\name{initial_distribution}
\alias{initial_distribution}
\title{inititial_distribution}
\usage{
initial_distribution(StSpM)
}
\arguments{
\item{StSpM}{matrix of all state}
}
\value{
an nrow(StSpM) probability with its first entry equal to 1
}
\description{
Computes the initial distribution in the DPH framework
}
