% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{KingmanCovMat25}
\alias{KingmanCovMat25}
\title{Kingman covariance matrix for n = 25}
\format{
A numeric matrix of dimension 253x253
}
\source{
Generated using covariance_matrix_F
}
\usage{
KingmanCovMat25
}
\description{
A precomputed covariance matrix under the Kingman coalescent
for ranked unlabeled rooted bifurcating trees with 25 leaves.
}
\details{
This matrix is used for statistical inference and hypothesis testing
in the RankedCoalescent package.

The matrix represents the covariances between all non-fixed F-matrix entries

It was precomputed to avoid expensive computations during runtime.
}
\keyword{datasets}
