% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{KingmanFrechetMean25_2}
\alias{KingmanFrechetMean25_2}
\title{KingmanFrechetMean25_2}
\format{
A matrix of dimension 24 x 24
}
\source{
Generated using ViTreebi
}
\usage{
KingmanFrechetMean25_2
}
\description{
The second of two Kingman Frechet mean F-matrices with 25 lineages
}
\details{
the matrix is the first of two Frechet means in the Kingman model
}
\keyword{datasets}
