% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{KingmanMeanNoFixed25}
\alias{KingmanMeanNoFixed25}
\title{Kingman mean of non fixed entries for n = 25
#' A precomputed mean vector under the Kingman coalescent
for ranked unlabeled rooted bifurcating trees with 25 leaves.}
\format{
A vector of length 253
}
\source{
Generated using mean_tree_DPH_nofixed
}
\usage{
KingmanMeanNoFixed25
}
\description{
This vector is used for statistical inference and hypothesis testing
in the RankedCoalescent package.
}
\details{
The vector represents the expectation all non-fixed F-matrix entries

It was precomputed to avoid expensive computations during runtime.
}
\keyword{datasets}
