% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityAndPathFunctions.R
\name{make_cost_tiers}
\alias{make_cost_tiers}
\title{make_cost_tiers}
\usage{
make_cost_tiers(Tiers, StSpM, MeanTree)
}
\arguments{
\item{Tiers}{A list of tiers made using make_tiers}

\item{StSpM}{matrix of all states}

\item{MeanTree}{mean F-matrix from a given distribution}
}
\value{
a list of all costs in the same format is tiers
}
\description{
Computes c(s) for all states s
}
