% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityAndPathFunctions.R
\name{make_Tiers}
\alias{make_Tiers}
\title{make_tiers}
\usage{
make_Tiers(StSpM)
}
\arguments{
\item{StSpM}{matrix of all states}
}
\value{
A list of all n-1 tiers
}
\description{
Puts states into tiers
}
