% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPHFunctions.R
\name{max_E}
\alias{max_E}
\title{max_E}
\usage{
max_E(n)
}
\arguments{
\item{n}{the number of leaves}
}
\value{
maximum external branch length
}
\description{
Computes the minimum sum of non-fixed entries
}
