% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPHFunctions.R
\name{max_S}
\alias{max_S}
\title{max_S}
\usage{
max_S(n)
}
\arguments{
\item{n}{the number of leaves}
}
\value{
maximum sum of non fixed entries
}
\description{
Computes the minimum sum of non-fixed entries
}
