% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPHFunctions.R
\name{mean_E}
\alias{mean_E}
\title{mean_E}
\usage{
mean_E(MeanTree)
}
\arguments{
\item{MeanTree}{mean F-matrix}
}
\value{
the mean external branch length
}
\description{
computes the mean external branch length
}
