% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPHFunctions.R
\name{mean_S}
\alias{mean_S}
\title{mean_S}
\usage{
mean_S(MeanTreeNoFixed)
}
\arguments{
\item{MeanTreeNoFixed}{vector of means of all non fixed entries}
}
\value{
the mean of the sum of non fixed entries
}
\description{
computes the mean of the sum of non fixed entries
}
