% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPHFunctions.R
\name{mean_tree_DPH}
\alias{mean_tree_DPH}
\title{mean_tree_DPH}
\usage{
mean_tree_DPH(n, LeftProducts, IndexList)
}
\arguments{
\item{n}{the number of leaves/lineages.}

\item{LeftProducts}{A list of the products init_dist * U * diag(r_{i,j})
where (i,j) is a non-fixed index. Computed via left_products_seq}

\item{IndexList}{A list of all non fixed indices. Made using \code{non_fixed_indices}}
}
\value{
The mean F matrix of the distribution with TPM used to compute
\code{LeftProducts}.
}
\description{
Computes the mean F-matrix with \verb{M_\{ij\} = E[F_\{ij\}]} where the columns of
\code{F} are sampled from the ranked coalescent with transition probabilities
used in \code{LeftProducts}.
}
