% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPHFunctions.R
\name{mean_tree_DPH_nofixed}
\alias{mean_tree_DPH_nofixed}
\title{mean_tree_DPH_nofixed}
\usage{
mean_tree_DPH_nofixed(LeftProducts, IndexList)
}
\arguments{
\item{LeftProducts}{A list of the products init_dist * U * diag(r_{i,j})
where (i,j) is a non-fixed index. Computed via left_products_seq}

\item{IndexList}{A list of all non fixed indices. Made using non_fixed_indices}
}
\value{
The mean F matrix of the distribution with TPM used to compute
LeftProducts.
}
\description{
Computes the non fixed entries of the mean F-matrix with \verb{M_\{ij\} = E[F_\{ij\}]}
where the columns of \code{F} are sampled from the ranked coalescent with
transition probabilities used in \code{LeftProducts}.
}
