% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityAndPathFunctions.R
\name{meanF_from_sample}
\alias{meanF_from_sample}
\title{meanF_from_sample}
\usage{
meanF_from_sample(Flist)
}
\arguments{
\item{Flist}{A list of F-matrices}
}
\value{
The mean F matrix of the sample
}
\description{
computes the mean F-matrix of a sample
}
