% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{MeanVarCov_SE}
\alias{MeanVarCov_SE}
\title{MeanVarCov_SE}
\format{
A data frame with columns n, MeanS, MeanE, VarS, VarE, CovSE
}
\source{
Generated using covariance_matrix_SE, mean_S, and mean_E
}
\usage{
MeanVarCov_SE
}
\description{
The mean, variance and covariance of external branch length and sum of
non-fixed entries for n between 4-25.
}
\details{
the data frame has 22 columns and 6 rows.
}
\keyword{datasets}
