% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityAndPathFunctions.R
\name{mergRep}
\alias{mergRep}
\title{mergeRep}
\usage{
mergRep(a, b)
}
\arguments{
\item{a}{column in which the process is}

\item{b}{column into which the process might transition}
}
\value{
Vector of differences, all but the last element reversed
}
\description{
Difference operator returning a low dimensional summary of two vectors coalescence potential
}
