% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPHFunctions.R
\name{min_E}
\alias{min_E}
\title{min_E}
\usage{
min_E(n)
}
\arguments{
\item{n}{the number of leaves}
}
\value{
minimum external branch length
}
\description{
Computes the minimum sum of non-fixed entries
}
