% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPHFunctions.R
\name{min_S}
\alias{min_S}
\title{min_S}
\usage{
min_S(n)
}
\arguments{
\item{n}{the number of leaves}
}
\value{
minimum sum of non fixed entries
}
\description{
Computes the minimum sum of non-fixed entries
}
