% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetupPreloads.R
\name{ndiff25_from_binM25}
\alias{ndiff25_from_binM25}
\title{ndiff25_from_binM25}
\usage{
ndiff25_from_binM25(binM25)
}
\arguments{
\item{binM25}{a matrix of all binary numbers that can be represented
with 25 digits. Computed using binary_matrix(25). Alternatively can be loaded.}
}
\value{
the ndiff25 vector.
}
\description{
Function computes the ndiffM25 vector from the binM25 matrix.
}
