% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityAndPathFunctions.R
\name{non_fixed_indices}
\alias{non_fixed_indices}
\title{non_fixed_indices}
\usage{
non_fixed_indices(n)
}
\arguments{
\item{n}{the number of leaves / lineages in the tree}
}
\value{
A list of all the non fixed indices
}
\description{
computes all the non-fixed entries and returns it as a list
}
