% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityAndPathFunctions.R
\name{path_to_Fmat}
\alias{path_to_Fmat}
\title{path_to_Fmat}
\usage{
path_to_Fmat(path, StSpM)
}
\arguments{
\item{path}{a vector of length n-1}

\item{StSpM}{matrix of all states}
}
\value{
the F-matrix corresponding to the path
}
\description{
Translates a path to its corresponding F-matrix
}
