% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_embedding.R
\name{ranked_coal}
\alias{ranked_coal}
\title{ranked_coal}
\usage{
ranked_coal(n, prob = TRUE)
}
\arguments{
\item{n}{Number from 3 to 23. The number of lineages at the start of the
process}

\item{prob}{If false, returns a matrix where possible transitions are 1 and
impossible ones are 0 instead of a transition probability matrix
(this is used for the ViTreebi)}
}
\value{
A list where the first element is the state space matrix, the
second is the transition probability/possibility matrix
}
\description{
Generates state space and transition probability matrix for given n
}
