% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_embedding.R
\name{ranked_coal_list}
\alias{ranked_coal_list}
\title{ranked_coal_list}
\usage{
ranked_coal_list(n, ChildAndProb = FALSE, TierInd = FALSE)
}
\arguments{
\item{n}{Number from 3 to 25. The number of lineages at the start of the process}

\item{ChildAndProb}{If true output contains two lists. childrenList is
a list of length = #states - n-2, where each entry is the possible next
states for that state. probList is a list containing the corresponding
transition probabilities.}

\item{TierInd}{if TRUE returns a tier list (import for ranked_coal_seq)}
}
\value{
A list where the first element is a state space list, followed by
lists of antecedents,children,probabilities (in the same order as children),
and the tier of each state
}
\description{
Same process as RankedCoal, but works for larger n by returning linked
lists instead of a matrix
}
