% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_embedding.R
\name{ranked_coal_seq}
\alias{ranked_coal_seq}
\title{ranked_coal_seq}
\usage{
ranked_coal_seq(n)
}
\arguments{
\item{n}{The number of lineages at the start of the process}
}
\value{
A list containing the state space matrix, the TierInd, and the
transition probability matrix in list format.
}
\description{
Generates state space matrix and transition probability matrix for given n
the transition probability matrix is generated as a list, where the k'th
entry contains the transition probability matrix from tier (k-1) to tier k.
}
