% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_embedding.R
\name{ranked_coal_StSpM_only}
\alias{ranked_coal_StSpM_only}
\title{ranked_coal_StSpM_only}
\usage{
ranked_coal_StSpM_only(n)
}
\arguments{
\item{n}{The number of lineages at the start of the process}
}
\value{
State space matrix of ranked unlabeled trees for n lineages
}
\description{
Generates state space matrix for given n
}
