% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPHFunctions.R
\name{reward_E_BCP}
\alias{reward_E_BCP}
\title{reward_E_BCP}
\usage{
reward_E_BCP(StSpMBCP)
}
\arguments{
\item{StSpMBCP}{State space matrix of the ranked BCP}
}
\value{
The reward vector of the external branch length
}
\description{
Computes the reward transform for E based on the BCP
}
