% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPHFunctions.R
\name{reward_matrix_F}
\alias{reward_matrix_F}
\title{reward_matrix_F}
\usage{
reward_matrix_F(StSpM, Tiers, IndexList)
}
\arguments{
\item{StSpM}{matrix of all states}

\item{Tiers}{A list of tiers made using make_tiers.}

\item{IndexList}{A list of all non fixed indices. Made using non_fixed_indices}
}
\value{
A nrow(StSpM) x (ncol(StSpM)-1)*(ncol(StSpM)-2)/2 reward matrix
}
\description{
Computes the reward matrix of the non fixed F-matrix entries from
equation (23) in NAME OF PAPER
}
