% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPHFunctions.R
\name{right_products_seq}
\alias{right_products_seq}
\title{right_products_seq}
\usage{
right_products_seq(SeqProbs, RewardMatrixF, Tiers, IndexList)
}
\arguments{
\item{SeqProbs}{List of transition probability matrix between consecutive
tiers. The k'th entry is equal to T_{k-1,k}, k = 1,...,n-2. Can be computed
using seq_ranked_coalescent.}

\item{RewardMatrixF}{reward matrix of the non-fixed entries. Can be computed
using reward_matrix_F.}

\item{Tiers}{A list of tiers made using make_tiers.}

\item{IndexList}{A list of all non fixed indices. Made using non_fixed_indices}
}
\value{
A list of all the products U*diag(r_{i,j})*e where i,j is a non
fixed index.
}
\description{
Computes the right factors U * diag(r_{i,j}) * e for all non fixed entries
i,j.
}
