% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleIslandModel.R
\name{rRankedIslandCoal}
\alias{rRankedIslandCoal}
\title{rRankedIslandCoal}
\usage{
rRankedIslandCoal(
  m,
  n,
  K,
  migrationRate = 1,
  coalescenceRate = 1,
  startDist = NA
)
}
\arguments{
\item{m}{Number of trees to simulate}

\item{n}{Number of lineages at start}

\item{K}{number of different islands}

\item{migrationRate}{Rate of migration}

\item{coalescenceRate}{Rate of coalescence}

\item{startDist}{If there is a fixed starting distribution, it can be given
using a vector of length K that sums to n. Otherwise, a random starting
distribution will be drawn using the multinomial distribution}
}
\value{
A list of n ranked unlabeled trees that were drawn according to the specifications of the parameters
}
\description{
Simulates from the Ranked Structured Coalescent
}
