% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_embedding.R
\name{seq_probM_to_probM}
\alias{seq_probM_to_probM}
\title{seq_probM_to_probM}
\usage{
seq_probM_to_probM(seqProb)
}
\arguments{
\item{seqProb}{a list of transition probability matrices between consecutive
tiers.}
}
\value{
An (Fib(n+1)-1) x (Fib(n+1) - 1) transition probability matrix
}
\description{
Functions returns a transition probability matrix given a list of transition
probability matrices between consecutive tiers.
}
