% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{StSpM25}
\alias{StSpM25}
\title{StSpM25}
\format{
A matrix with 121,392 rows and 24 columns
}
\source{
Generated using the ranked_coalescent function
}
\usage{
StSpM25
}
\description{
The state space of the Markov embedding with n = 25 lineages.
}
\details{
Each row is a state. The state represents a column in an Fmatrix.
}
\keyword{datasets}
