% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityAndPathFunctions.R
\name{sum_nonfixed_entries}
\alias{sum_nonfixed_entries}
\title{sum_nonfixed_entries}
\usage{
sum_nonfixed_entries(Fmat)
}
\arguments{
\item{Fmat}{An (n-1) x (n-1) F-matrix}
}
\value{
the sum of non fixed entries
}
\description{
computes the sum of non fixed entries in an F-matrix
}
