% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityAndPathFunctions.R
\name{toBool}
\alias{toBool}
\title{toBool}
\usage{
toBool(val)
}
\arguments{
\item{val}{some 1 dimensional object}
}
\value{
TRUE or FALSE, FALSE only if val is some form of FALSE/0
}
\description{
Ensures that some input value is boolean
}
