% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPHFunctions.R
\name{var_E}
\alias{var_E}
\title{var_E}
\usage{
var_E(CovMatNoFixed, IndexList)
}
\arguments{
\item{CovMatNoFixed}{the covariance matrix of all non fixed entries}

\item{IndexList}{A list of all non fixed indices. Made using non_fixed_indices}
}
\value{
the variance of the external branch length
}
\description{
computes the variance of the external branch length
}
