% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPHFunctions.R
\name{var_S}
\alias{var_S}
\title{var_S}
\usage{
var_S(CovMatNoFixed)
}
\arguments{
\item{CovMatNoFixed}{the covariance matrix of all non fixed entries}
}
\value{
the variance of the sum of the non fixed entries
}
\description{
computes the variance of the sum of non fixed entries
}
