% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViTreebi.R
\name{ViTreebi}
\alias{ViTreebi}
\title{ViTreebi}
\usage{
ViTreebi(n, AntecedentList, StSpM, CostTiers, Tiers)
}
\arguments{
\item{n}{the number of leaves / lineages in the trees}

\item{AntecedentList}{a list of length Fib(n+1)-1. Each entry contains the
possible previous states. Can be computed using BigRankedCoal}

\item{StSpM}{matrix of all states}

\item{CostTiers}{A list of costs for each state. Made using make_cost_tiers}

\item{Tiers}{A list of tiers made using make_tiers.}
}
\value{
all Frechet mean paths as well as the minimal attainable cost
}
\description{
Computes the minimum attainable distance to the min tree and the paths
attaining this minimum.
}
