% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticalTests.R
\name{wald_test_F}
\alias{wald_test_F}
\title{wald_test_F}
\usage{
wald_test_F(Fmats, M0, Sigma0invSqrt, IndexList)
}
\arguments{
\item{Fmats}{a list of (n-1) x (n-1) F-matrices}

\item{M0}{the mean F-matrix under the null hypothesis. Can be
computed using mean_tree_DPH_nofixed}

\item{Sigma0invSqrt}{The inverse square root of the covariance matrix of
all non fixed entries under the null hypothesis. Can be computed using
covariance_matrix_F}

\item{IndexList}{A list of all non fixed indices.
Made using \code{non_fixed_indices}}
}
\value{
the p-value of the test
}
\description{
Performs the wald test based on all non-fixed entries as described in
Theorem 6.1 in NAME OF PAPER.
}
