% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticalTests.R
\name{wald_test_SE}
\alias{wald_test_SE}
\title{wald_test_SE}
\usage{
wald_test_SE(Fmats, Mu0SE, Sigma0SEinvSqrt, IndexList)
}
\arguments{
\item{Fmats}{a list of (n-1) x (n-1) F-matrices}

\item{Mu0SE}{a vector of mean(S), mean(E) under the null hypothesis. Can be
computed using mean_E and mean_S}

\item{Sigma0SEinvSqrt}{The inverse square root of the covariance matrix
of E and S under the null hypothesis. Compute using covariance_matrix_SE}

\item{IndexList}{A list of all non fixed indices.
Made using \code{non_fixed_indices}}
}
\value{
the p-value of the test
}
\description{
Performs the wald test based on the balance indices S, E as described in
Theorem 6.1 in NAME OF PAPER.
}
